refactor(xenperf): set a multi-arch source structure
This version has only been tested for xen 4.4
The actual arch-dependent sources are now selected during the
compilation and only the matching between compilation arch and
runtime arch is tested.
The performance counters are now activated in a while loop and reset
at each iteration to avoid overflows.
It is now possible to specify a core on which to run an perfcnt
function.