fix(xenperf): detect pmc corruption in perfcnt_read
Before to read the MSR in perfcnt_read, the harware event selection
mask is read and compared to the expected one.
If their differ, a corruption is detected, the expected mask is
written again and an error code is returned.
Also remove the bitsize field in legacy_perfcnt for AMD/16/ since it
is a constant.