fix(toolbox): modify the wait condiftion of xen-shutdown
For DomU with large amount of memory, after the DomU is shutted down,
its name become "(null)" but Xen cannot free its memory
instantaneously.
Before, xen-shutdown was waiting to not find the sutted down vm name
in xl list and could allow to continue whereas all the memory was not
freed.
Now it wait for the ID of the DomU to disappear (the ID is kept until
the DomU truely disappear).