fix(toolbox): fix some compilation bug in build-dc
Fix a ld compilation bug for build-dc: the -lgomp option was not
placed at the end of the command and so was ignored on some version
of gcc.
Now it is placed right before the '-lm' option, so it should work for
every version of gcc.
Also make the compilation quiet, especially for gcc version dependent
warnings.